Extract SdStateManager from server mod
